+++ frontend/src/app/modules/page/other-parts/presentation/supplier-parts/supplier-parts.component.ts
@@ -21,7 +21,7 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { Pagination } from '@core/model/pagination.model';
 import { OtherPartsFacade } from '@page/other-parts/core/other-parts.facade';
-import { Part } from '@page/parts/model/parts.model';
+import { Part, SemanticDataModel } from '@page/parts/model/parts.model';
 import { CreateHeaderFromColumns, TableConfig, TableEventConfig } from '@shared/components/table/table.model';
 import { View } from '@shared/model/view.model';
 import { PartDetailsFacade } from '@shared/modules/part-details/core/partDetails.facade';
@@ -35,6 +35,7 @@ import { BehaviorSubject, Observable, Subject } from 'rxjs';
 export class SupplierPartsComponent implements OnInit, OnDestroy {
   public readonly displayedColumns: string[] = [
     'select',
+    'semanticDataModel',
     'name',
     'manufacturer',
     'partNumber',
@@ -50,6 +51,7 @@ export class SupplierPartsComponent implements OnInit, OnDestroy {
     serialNumber: true,
     batchNumber: true,
     productionDate: true,
+    semanticDataModel: true,
   };
 
   public readonly tableConfig: TableConfig = {
@@ -77,6 +79,10 @@ export class SupplierPartsComponent implements OnInit, OnDestroy {
   }
 
   public get currentSelectedItems(): Part[] {
+
+    this.selectedItems = this.selectedItems.map(part => {
+      return {...part, semanticDataModel: SemanticDataModel[part.semanticDataModel.toUpperCase() as keyof typeof SemanticDataModel]}
+    })
     return this.selectedItems || [];
   }