+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 - Updated spring boot to 3.0.7 to fix: CVE-2023-20883
+- Fixed calculation of otherParts amount within dashboard response
 
 ### Removed
 
@@ -26,6 +27,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Cors configuration for environment e2e-a and e2e-b
 - GitHubAction to update dependencies files on merge to main
 - Added testdata with batches
+- Added Semantic Data Model to frontend to support batches
 - Added batch aspect to the asset import process
 
 ### Changed